Fix compilation on Mac OS X.
In `vendor/libxml/include/libxml/xmlversion.h`, LIBXML_THREAD_ENABLED is only
enabled if there's an appropriate "I want threads" macro defined. None of the
appropriate macros appear to be defined by default on OS X, so define
`_REENTRANT` explicitly in libxml.gyp.

In threads.c: PTHREAD_ONCE_INIT is a static initializer according the the POSIX
threads specification, but happens to be `0` on glibc so you can get away with
using it outside of a static initialization context. It expands to
`{_PTHREAD_ONCE_SIG_init, {0}}` on OS X, hence the error.

There may be prettier ways to do one or both of these, but the fixes appear to
work on this 10.9.5 machine, and the test suite passes. Currently untested on
10.10, but I suspect @alxndrsn will be able to give this a test soon.